+++ src/modules/interfaces/react/tests/integration/operation-stream-simulation.js
@@ -48,7 +48,6 @@ function ensureTestConfig() {
     memoryMode: 'operation',
     keepMemory: true,
     outputDir: './outputs',
-    unifiedOutput: true,
     theme: 'retro',
     showMemoryUsage: false,
     showOperationId: true,